+++ lib/reflection_factory.dart
@@ -9,3 +9,4 @@ export 'src/reflection_factory_annotation.dart';
 export 'src/reflection_factory_base.dart';
 export 'src/reflection_factory_extension.dart';
 export 'src/reflection_factory_json.dart';
+export 'src/reflection_factory_type.dart';