+++ test/reflection_factory_test.dart
@@ -580,6 +580,31 @@ void main() {
               .invoke([1.1]),
           isTrue);
 
+      var domainReflection = TestDomainWithReflection$reflection.staticInstance;
+
+      expect(
+          domainReflection.fieldsNames, equals(['hashCode', 'name', 'suffix']));
+      expect(domainReflection.methodsNames, equals(['toJson', 'toString']));
+      expect(
+          domainReflection.constructorsNames, equals(['', 'named', 'parse']));
+
+      {
+        var constructor = domainReflection.constructor('named')!;
+
+        expect(constructor.allParametersNames, equals(['name', 'suffix']));
+
+        var parameterName = constructor.namedParameters['name']!;
+        var parameterSuffix = constructor.namedParameters['suffix']!;
+
+        expect(parameterName.required, isTrue);
+        expect(parameterName.nullable, isFalse);
+        expect(parameterName.type, equals(TypeReflection.tString));
+
+        expect(parameterSuffix.required, isFalse);
+        expect(parameterSuffix.nullable, isFalse);
+        expect(parameterSuffix.type, equals(TypeReflection.tString));
+      }
+
       var opReflection = TestOpWithReflection$reflection.staticInstance;
       var opAReflection = TestOpAWithReflection$reflection.staticInstance;
       var opBReflection = TestOpBWithReflection$reflection.staticInstance;
@@ -588,9 +613,9 @@ void main() {
       expect(opAReflection.supperTypes, equals([TestOpWithReflection]));
       expect(opBReflection.supperTypes, equals([TestOpWithReflection]));
 
-      expect(opReflection.fieldsNames, equals(['type']));
+      expect(opReflection.fieldsNames, equals(['type', 'value']));
       expect(opAReflection.fieldsNames, equals(['type', 'value']));
-      expect(opBReflection.fieldsNames, equals(['amount', 'type']));
+      expect(opBReflection.fieldsNames, equals(['amount', 'type', 'value']));
 
       expect(
           opAReflection